# backend/api/src/main/python/openapi_server/controllers/auth_controller.py
import connexion
from typing import Dict
from typing import Tuple
from typing import Union

from openapi_server import util
from openapi_server.models.auth_request import AuthRequest
from openapi_server.models.auth_response import AuthResponse
from openapi_server.models.password_reset_request import PasswordResetRequest

def auth_logout_post():  # noqa: E501
    """Logout current user (invalidate token/session)

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]]
    """
    # CMZ Auto-Generated Implementation Connection
    # This template automatically connects controllers to impl modules
    try:
        # Dynamic import of implementation module based on controller name
        # Auto-detect implementation module from operationId
        impl_module_name = "authcontroller".replace("_controller", "")
        impl_function_name = "handle_"

        # Try common implementation patterns
        try:
            # Pattern 1: Direct module import
            impl_module = __import__(f"openapi_server.impl.{impl_module_name}", fromlist=[impl_function_name])
            impl_function = getattr(impl_module, impl_function_name)
        except (ImportError, AttributeError):
            # Pattern 2: Generic handler with hexagonal architecture routing
            from openapi_server.impl import handlers
            # Use the generic handle_ function that routes based on caller name
            impl_function = handlers.handle_
            if not impl_function:
                # Pattern 3: Default error for missing implementation
                raise NotImplementedError(f"Implementation function 'handle_' not found in handlers module")

        # Call implementation function with processed parameters
        result = impl_function()

        # Handle different return types
        if isinstance(result, tuple):
            return result  # Already formatted (data, status_code)
        else:
            return result, 204

    except NotImplementedError as e:
        # Development mode: return clear error instead of placeholder
        from openapi_server.models.error import Error
        error_obj = Error(
            code="not_implemented",
            message=f"Controller auth_logout_post implementation not found: {str(e)}",
            details={"controller": "AuthController", "operation": "auth_logout_post"}
        )
        return error_obj, 501

    except Exception as e:
        # Use centralized error handler if available
        try:
            from openapi_server.impl.error_handler import handle_exception_for_controllers
            return handle_exception_for_controllers(e)
        except ImportError:
            # Fallback error response
            from openapi_server.models.error import Error
            error_obj = Error(
                code="internal_error",
                message=f"Internal server error in auth_logout_post: {str(e)}",
                details={"controller": "AuthController", "operation": "auth_logout_post"}
            )
            return error_obj, 500


def auth_post():  # noqa: E501
    """Login or register

    :param auth_request:
    :type auth_request:  | bytes

    :rtype: Union[AuthResponse, Tuple[AuthResponse, int], Tuple[AuthResponse, int, Dict[str, str]]]
    """
    # Auto-generated parameter handling
    if connexion.request.is_json:
        auth_request = AuthRequest.from_dict(connexion.request.get_json())  # noqa: E501

    # CMZ Auto-Generated Implementation Connection
    # This template automatically connects controllers to impl modules
    try:
        # Dynamic import of implementation module based on controller name
        # Auto-detect implementation module from operationId
        impl_module_name = "authcontroller".replace("_controller", "")
        impl_function_name = "handle_"

        # Try common implementation patterns
        try:
            # Pattern 1: Direct module import
            impl_module = __import__(f"openapi_server.impl.{impl_module_name}", fromlist=[impl_function_name])
            impl_function = getattr(impl_module, impl_function_name)
        except (ImportError, AttributeError):
            # Pattern 2: Generic handler with hexagonal architecture routing
            from openapi_server.impl import handlers
            # Use the generic handle_ function that routes based in caller name
            impl_function = handlers.handle_
            if not impl_function:
                # Pattern 3: Default error for missing implementation
                raise NotImplementedError(f"Implementation function 'handle_' not found in handlers module")

        # Call implementation function with processed parameters
        # The handlers expect the raw JSON body
        body = connexion.request.get_json()
        result = impl_function(body)

        # Handle different return types
        if isinstance(result, tuple):
            return result  # Already formatted (data, status_code)
        else:
            return result, 200

    except NotImplementedError as e:
        # Development mode: return clear error instead of placeholder
        from openapi_server.models.error import Error
        error_obj = Error(
            code="not_implemented",
            message=f"Controller auth_post implementation not found: {str(e)}",
            details={"controller": "AuthController", "operation": "auth_post"}
        )
        return error_obj, 501

    except Exception as e:
        # Use centralized error handler if available
        try:
            from openapi_server.impl.error_handler import handle_exception_for_controllers
            return handle_exception_for_controllers(e)
        except ImportError:
            # Fallback error response
            from openapi_server.models.error import Error
            error_obj = Error(
                code="internal_error",
                message=f"Internal server error in auth_post: {str(e)}",
                details={"controller": "AuthController", "operation": "auth_post"}
            )
            return error_obj, 500


def auth_refresh_post():  # noqa: E501
    """Refresh access token

    :rtype: Union[AuthResponse, Tuple[AuthResponse, int], Tuple[AuthResponse, int, Dict[str, str]]]
    """
    # CMZ Auto-Generated Implementation Connection
    # This template automatically connects controllers to impl modules
    try:
        # Dynamic import of implementation module based on controller name
        # Auto-detect implementation module from operationId
        impl_module_name = "authcontroller".replace("_controller", "")
        impl_function_name = "handle_"

        # Try common implementation patterns
        try:
            # Pattern 1: Direct module import
            impl_module = __import__(f"openapi_server.impl.{impl_module_name}", fromlist=[impl_function_name])
            impl_function = getattr(impl_module, impl_function_name)
        except (ImportError, AttributeError):
            # Pattern 2: Generic handler with hexagonal architecture routing
            from openapi_server.impl import handlers
            # Use the generic handle_ function that routes based on caller name
            impl_function = handlers.handle_
            if not impl_function:
                # Pattern 3: Default error for missing implementation
                raise NotImplementedError(f"Implementation function 'handle_' not found in handlers module")

        # Call implementation function with processed parameters
        result = impl_function()

        # Handle different return types
        if isinstance(result, tuple):
            return result  # Already formatted (data, status_code)
        else:
            return result, 200

    except NotImplementedError as e:
        # Development mode: return clear error instead of placeholder
        from openapi_server.models.error import Error
        error_obj = Error(
            code="not_implemented",
            message=f"Controller auth_refresh_post implementation not found: {str(e)}",
            details={"controller": "AuthController", "operation": "auth_refresh_post"}
        )
        return error_obj, 501

    except Exception as e:
        # Use centralized error handler if available
        try:
            from openapi_server.impl.error_handler import handle_exception_for_controllers
            return handle_exception_for_controllers(e)
        except ImportError:
            # Fallback error response
            from openapi_server.models.error import Error
            error_obj = Error(
                code="internal_error",
                message=f"Internal server error in auth_refresh_post: {str(e)}",
                details={"controller": "AuthController", "operation": "auth_refresh_post"}
            )
            return error_obj, 500


def auth_reset_password_post(password_reset_request):  # noqa: E501
    """Initiate password reset

    :param password_reset_request: 
    :type password_reset_request:  | bytes

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]]
    """
    # Auto-generated parameter handling
    if connexion.request.is_json:
        password_reset_request = PasswordResetRequest.from_dict(connexion.request.get_json())  # noqa: E501

    # CMZ Auto-Generated Implementation Connection
    # This template automatically connects controllers to impl modules
    try:
        # Dynamic import of implementation module based on controller name
        # Auto-detect implementation module from operationId
        impl_module_name = "authcontroller".replace("_controller", "")
        impl_function_name = "handle_"

        # Try common implementation patterns
        try:
            # Pattern 1: Direct module import
            impl_module = __import__(f"openapi_server.impl.{impl_module_name}", fromlist=[impl_function_name])
            impl_function = getattr(impl_module, impl_function_name)
        except (ImportError, AttributeError):
            # Pattern 2: Generic handler with hexagonal architecture routing
            from openapi_server.impl import handlers
            # Use the generic handle_ function that routes based on caller name
            impl_function = handlers.handle_
            if not impl_function:
                # Pattern 3: Default error for missing implementation
                raise NotImplementedError(f"Implementation function 'handle_' not found in handlers module")

        # Call implementation function with processed parameters
        result = impl_function(password_reset_request)

        # Handle different return types
        if isinstance(result, tuple):
            return result  # Already formatted (data, status_code)
        else:
            return result, 204

    except NotImplementedError as e:
        # Development mode: return clear error instead of placeholder
        from openapi_server.models.error import Error
        error_obj = Error(
            code="not_implemented",
            message=f"Controller auth_reset_password_post implementation not found: {str(e)}",
            details={"controller": "AuthController", "operation": "auth_reset_password_post"}
        )
        return error_obj, 501

    except Exception as e:
        # Use centralized error handler if available
        try:
            from openapi_server.impl.error_handler import handle_exception_for_controllers
            return handle_exception_for_controllers(e)
        except ImportError:
            # Fallback error response
            from openapi_server.models.error import Error
            error_obj = Error(
                code="internal_error",
                message=f"Internal server error in auth_reset_password_post: {str(e)}",
                details={"controller": "AuthController", "operation": "auth_reset_password_post"}
            )
            return error_obj, 500