# backend/api/src/main/python/openapi_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  description: API for the Cougar Mountain Zoo digital ambassador platform
  title: CMZ API
  version: 1.0.2
servers:
- description: Production
  url: https://api.cmz.example.com
- description: Staging
  url: https://staging.api.cmz.example.com
- description: Local
  url: http://localhost:8080
...
paths:
...
  /auth:
    post:
      operationId: auth_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              example:
                token: eyJhbGciOi...
                expiresIn: 3600
                user:
                  userId: user_1
                  email: user1@cmz.org
                  displayName: User 1
                  role: admin
                  created:
                    at: 2025-08-11T09:00:00Z
                    by:
                      userId: user_1
                      email: system@cmz.org
                      displayName: System
                  modified:
                    at: 2025-08-11T09:00:00Z
                    by:
                      userId: user_1
                      email: system@cmz.org
                      displayName: System
                  softDelete: false
              schema:
                $ref: "#/components/schemas/AuthResponse"
          description: Auth token
        "401":
          $ref: "#/components/responses/Unauthorized"
      summary: Login or register
      tags:
      - Auth
      x-openapi-router-controller: openapi_server.controllers.auth_controller
...
  /auth/reset_password:
    post:
      operationId: auth_reset_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
        required: true
      responses:
        "204":
          description: Email sent if account exists
      summary: Initiate password reset
      tags:
      - Auth
      x-openapi-router-controller: openapi_server.controllers.auth_controller
...
components:
  responses:
    Unauthorized:
      content:
        application/json:
          examples:
            unauthorized:
              value:
                code: unauthorized
                message: Missing or invalid token
          schema:
            $ref: "#/components/schemas/Error"
      description: Unauthorized
...
  schemas:
    ActorRef:
      example:
        actorId: actorId
        displayName: displayName
        email: email
      properties:
        actorId:
          title: actorId
          type: string
        email:
          format: email
          title: email
          type: string
        displayName:
          title: displayName
          type: string
      title: ActorRef
      type: object
    AuditStamp:
      example:
        at: 2000-01-23T04:56:07.000+00:00
        by:
          actorId: actorId
          displayName: displayName
          email: email
      properties:
        at:
          format: date-time
          title: at
          type: string
        by:
          $ref: "#/components/schemas/ActorRef"
      title: AuditStamp
      type: object
    AuditFields:
      properties:
        created:
          $ref: "#/components/schemas/AuditStamp"
        modified:
          $ref: "#/components/schemas/AuditStamp"
        deleted:
          $ref: "#/components/schemas/AuditStamp"
        softDelete:
          default: false
          title: softDelete
          type: boolean
      title: AuditFields
      type: object
    Error:
      example:
        code: code
        details:
          key: ""
        message: message
      properties:
        code:
          title: code
          type: string
        message:
          title: message
          type: string
        details:
          additionalProperties: true
          title: details
          type: object
      required:
      - code
      - message
      title: Error
      type: object
...
    AuthRequest:
      example:
        password: password
        username: username
        register: true
      properties:
        username:
          format: email
          title: username
          type: string
        password:
          title: password
          type: string
        register:
          description: "If true, create account if not found"
          title: register
          type: boolean
      required:
      - password
      - username
      title: AuthRequest
      type: object
    AuthResponse:
      example:
        expiresIn: 0
        user:
          familyId: familyId
          phoneNumber: phoneNumber
          role: visitor
          deleted:
            at: 2000-01-23T04:56:07.000+00:00
            by:
              actorId: actorId
              displayName: displayName
              email: email
          softDelete: false
          displayName: displayName
          created:
            at: 2000-01-23T04:56:07.000+00:00
            by:
              actorId: actorId
              displayName: displayName
              email: email
          modified:
            at: 2000-01-23T04:56:07.000+00:00
            by:
              actorId: actorId
              displayName: displayName
              email: email
          userType: none
          userId: userId
          email: email
          age: 73
        token: token
      properties:
        token:
          title: token
          type: string
        expiresIn:
          description: Seconds until expiry
          title: expiresIn
          type: integer
        user:
          $ref: "#/components/schemas/User"
      required:
      - expiresIn
      - token
      - user
      title: AuthResponse
      type: object
    PasswordResetRequest:
      example:
        email: email
      properties:
        email:
          format: email
          title: email
          type: string
      required:
      - email
      title: PasswordResetRequest
      type: object
    User:
      allOf:
      - properties:
          userId:
            description: Unique user identifier
            pattern: "^[a-zA-Z0-9_-]+$"
            type: string
          email:
            description: RFC 5322 compliant email address
            format: email
            pattern: "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$"
            type: string
          displayName:
            description: Display name with basic character validation
            maxLength: 100
            minLength: 1
            pattern: "^[a-zA-Z0-9\\s\\-\\.']+$"
            type: string
          phoneNumber:
            description: E.164 international phone number format
            nullable: true
            pattern: "^\\+[1-9]\\d{1,14}$"
            type: string
          age:
            description: Age validation for reasonable ranges
            maximum: 120
            minimum: 3
            nullable: true
            type: integer
          role:
            enum:
            - visitor
            - user
            - member
            - editor
            - admin
            type: string
          userType:
            default: none
            enum:
            - none
            - parent
            - student
            type: string
          familyId:
            nullable: true
            type: string
        type: object
      - $ref: "#/components/schemas/AuditFields"
      example:
        familyId: familyId
        phoneNumber: phoneNumber
        role: visitor
        deleted:
          at: 2000-01-23T04:56:07.000+00:00
          by:
            actorId: actorId
            displayName: displayName
            email: email
        softDelete: false
        displayName: displayName
        created:
          at: 2000-01-23T04:56:07.000+00:00
          by:
            actorId: actorId
            displayName: displayName
            email: email
        modified:
          at: 2000-01-23T04:56:07.000+00:00
          by:
            actorId: actorId
            displayName: displayName
            email: email
        userType: none
        userId: userId
        email: email
        age: 73
      title: User
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: openapi_server.controllers.security_controller.info_from_bearerAuth